🧪 [testing improvement] Decouple and test sanitizeOffices
Refactored `sanitizeOffices` into a dedicated `api/_lib/sanitizer.ts` utility to facilitate unit testing without external API client side-effects.

🎯 **What:** Decoupled LLM output validation logic and implemented comprehensive unit tests.
📊 **Coverage:**
- Happy path validation and normalization.
- Missing field handling (city/country).
- Canonical region and officeType enforcement.
- City/Country deduplication.
- Input type safety and string trimming.
✨ **Result:** Improved testability and reliability of the office discovery pipeline. Unit tests now run in all environments without requiring the `openai` SDK to be present.
